requirements.txt
streamlit
langgraph
langchain-google-genai
langchain-tavily
langchain-community
langchain-core
python-dotenv
langgraph-checkpoint-sqlite
line-bot-sdk
psycopg2-binary
langgraph-checkpoint-postgres
psycopg[binary]
streamlit-authenticator